Feature network

+++ app/src/main/java/com/kristof/weather/models/Location.kt
@@ -1,6 +1,6 @@
 package com.kristof.weather.models
 
 data class Location(
-        var lat: Long,
-        var lon: Long
+        var lat: Double,
+        var lon: Double
 )
